Rust sdk eventlog support (#64)
* docs(contributor): contrib-readme-action has updated readme

* this commit enables get_cc_eventlog API:
- enhance cctrusted_base crate for eventlog parsing
- enhance cctrusted_vm crate for get_cc_eventlog API implementation
- UT cases for get_cc_eventlog
- add sample usage in sample crate

---------

+++ common/rust/cctrusted_base/src/tdx/rtmr.rs
@@ -3,11 +3,11 @@ use anyhow::anyhow;
 
 pub struct TdxRTMR {
     index: u8,
-    digest: (u8, TcgDigest),
+    digest: (u16, TcgDigest),
 }
 
 impl TdxRTMR {
-    pub fn new(index: u8, algo_id: u8, digest: [u8; 48]) -> Result<TdxRTMR, anyhow::Error> {
+    pub fn new(index: u8, algo_id: u16, digest: [u8; 48]) -> Result<TdxRTMR, anyhow::Error> {
         match TdxRTMR::is_valid_index(index) {
             Ok(_) => (),
             Err(e) => return Err(anyhow!("error creating TdxRTMR {:?}", e)),
@@ -39,7 +39,7 @@ impl TcgIMR for TdxRTMR {
         self.index
     }
 
-    fn get_tcg_digest(&self, _algo_id: u8) -> TcgDigest {
+    fn get_tcg_digest(&self, _algo_id: u16) -> TcgDigest {
         self.digest.1.clone()
     }
 
@@ -51,7 +51,7 @@ impl TcgIMR for TdxRTMR {
         Ok(true)
     }
 
-    fn is_valid_algo(algo_id: u8) -> Result<bool, anyhow::Error> {
+    fn is_valid_algo(algo_id: u16) -> Result<bool, anyhow::Error> {
         if algo_id != TPM_ALG_SHA384 {
             return Err(anyhow!("[is_valid_algo] invalid algo id: {}", algo_id));
         }
